fixed percentage
