+++ CalculatorApp/.idea/shelf/Uncommitted_changes_before_rebase__Changes_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_rebase_[Changes]" date="1710526101991" recycled="true" deleted="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_rebase_[Changes]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before rebase [Changes]" />
+</changelist>
