+++ CalculatorApp/app/src/main/java/com/thalajaat/calyxcalculator/utils/DeleteAllLongClickListener.kt
@@ -4,7 +4,7 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.widget.Toast
-import com.thalajaat.calyxcalculator.dormain.CalculationHandler
+import com.thalajaat.calyxcalculator.domain.CalculationHandler
 import com.thalajaat.calyxcalculator.utils.Utils.CLEAR_ALL_CLICK_ACTION
